Add fully functional downward arrow technique
-Downward arrow technique now appears by default if there is any text
in any of the three fields and also appears when the reveal button is
pushed. No weird spaces required, no cursor leaving the box when you
type the first character
-Note that if the user saves, leaves the route, returns, and then
deletes all the content in a given layer, that layer will disappear,
which can result in users having hidden layers. This isn't desirable
but is ultimately an uncommon and not particularly problematic bug
so it is a low priority fix.