Add formatting adjustments
-Center the various "view all" headings and contents. As a general rule,
input fields remained left-justified while text was center-justified.
-Changed the generics update notifier to show on clicking the save
button rather than on successful update, as the loop caused the
notification to appear as many times as there were entries. Note that
while this is a superior UI experience it is now inaccurate and may
indicate saving when saving has not been successful.